RISC-V: Refactor the testcases for rvv binop and cmp
This patch would like to refactor the testcases for rvv binop
and cmp after sorts of optimization option passing to testcase.
To fits different optimization option asm dump checks.

The below test suites are passed for this patch.
* The rv64gcv fully regression test.

It is test only patch and obvious up to a point, will commit it
directly if no comments in next 48H.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/autovec/binop/vcompress-avlprop-1.c: Skip
	m8 as it has different body layout.
	* gcc.target/riscv/rvv/autovec/cmp/cmp_vi-1.c: Add build option
	condition when check asm dumps.
	* gcc.target/riscv/rvv/autovec/cmp/cmp_vi-2.c: Ditto.
	* gcc.target/riscv/rvv/autovec/cmp/cmp_vi-5.c: Ditto.
	* gcc.target/riscv/rvv/autovec/cmp/cmp_vi-6.c: Ditto.
	* gcc.target/riscv/rvv/autovec/cmp/cmp_vi-9.c: Ditto.

Signed-off-by: Pan Li <pan2.li@intel.com>