ada: Remove workaround for RM_Size being unable to represent unknown size
When validating instances of Ada.Unchecked_Conversion, we explicitly detected
null arrays, because a size of 0 was used to indicate both an unknown size and
an actual size of 0. This limitation has been lifted, so we can remove
detection of null arrays.

Code cleanup; behavior is unaffected.

gcc/ada/ChangeLog:

	* sem_ch13.adb (Validate_Unchecked_Conversions): Remove detection of
	null arrays; remove tests for sizes being present, which are redundant
	after calling Known_Static_RM_Size.