final: Fix call to INSN_LOCATION on a NOTE rtl
This function has a code path that calls INSN_LOCATION on an rtl note. For a
note, this returns the note type enum rather than a location, but it runs
without complaint even with --enable-checking=rtl because both are stored in
the rt_int member of the rtunion. A subsequent commit will add a new rtl
format code specifically for locations, in which case attempting to call
INSN_LOCATION on a note will trigger an error. Fix it up by handling the
case of a note missing a location separately.

gcc/ChangeLog:

	* final.cc (reemit_insn_block_notes): Don't try to call
	INSN_LOCATION on a NOTE rtl object. Don't call change_scope () for a
	NOTE missing a location.